requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-cognitiveservices-speech==1.36.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
ffmpeg-python==0.2.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
influxdb-client==1.41.0
joblib==1.4.2
libusb1==3.1.0
llama-index==0.10.43
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.43
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
lz4==4.3.3
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
ndeflib==0.3.3
nest-asyncio==1.6.0
networkx==3.3
nfcpy==1.0.4
nltk==3.8.1
numpy==1.26.4
openai==1.14.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
protobuf==4.25.3
protoc-gen-openapiv2==0.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pyDes==2.0.1
pydub==0.25.1
pymongo==4.7.3
pypdf==4.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
reactivex==4.0.4
regex==2024.5.15
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
striprtf==0.0.26
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
